+++ lib/src/proxy/_.dart
@@ -31,6 +31,11 @@ js.JsObject _jsAnnotation(Annotation a) => _jsMap({
   'type': a.type
 });
 
+js.JsObject _jsBindKey(BindKey bk) => _jsMap({
+  'win' : bk.win,
+  'mac' : bk.mac
+});
+
 js.JsObject _jsDelta(d) {
   final jsMap = _jsMap({ 'action': d.action, 'range': _jsRange(d.range) });
   if (d.action == 'insertLines' || d.action == 'removeLines') {
@@ -82,6 +87,8 @@ Annotation _annotation(proxy) => new Annotation(
     row : proxy['row'] == null ? 0 : proxy['row'],
     text: proxy['text'],
     type: proxy['type'] == null ? Annotation.INFO : proxy['type']);
+
+BindKey _bindKey(proxy) => new BindKey(mac: proxy['mac'], win: proxy['win']);
 
 Delta _delta(proxy) {
   switch(proxy['action']) {