+++ test/src/editor_test.dart
@@ -664,3 +664,25 @@ void testPaste() {
   editor.paste('42');
   expect(editor.value, equals('snarf42'));
 }
+
+@Test()
+void testGetCommandManager() {
+  final cm = editor.commands;
+  expect(cm, isNotNull);
+  final commands = cm.getCommands();
+  expect(commands, isNot(isEmpty));
+}
+
+@Test()
+void testExecCommand() {
+  final c = new Command(
+      'paste-answer',
+      const BindKey(mac: 'Command-A', win: 'Ctrl-A'),
+      (editor) {
+        editor.paste('42');
+      });  
+  editor.setValue('snarf', 1);
+  editor.commands.addCommand(c);
+  editor.execCommand('paste-answer');
+  expect(editor.value, equals('snarf42'));
+}